DatLA update code from dev

+++ C0722G1-Repo-BE/src/main/java/com/c0722g1repobe/dto/notification/NotificationAllPropertyDto.java
@@ -6,7 +6,7 @@ package com.c0722g1repobe.dto.notification;
  * For: to show list notification by interface with all property
  */
 public interface NotificationAllPropertyDto {
-    Integer getIdNotification();
+    Long getIdNotification();
     String getPostingDate();
     String getTitle();
     String getContent();
