+++ C0722G1-Repo-BE/src/main/java/com/c0722g1repobe/dto/notification/NotificationDeleteDto.java
@@ -6,6 +6,6 @@ package com.c0722g1repobe.dto.notification;
  * For: to show list notification to delete
  */
 public interface NotificationDeleteDto {
-    Integer getIdNotification();
+    Long getIdNotification();
     String getTitle();
 }
